chore(core): fix according to comments
